Remove smoke.py and clean up base test classes
This commit removes the smoke.py file and reworks the classes in
test.py to be a bit simpler. Previously, DefaultClientTest
was only really needed by smoke.py for the smoke tests. (the only
place where python clients were used) However, it remained a separate
class because of the status_timeout() method. This breaks out that
method and removes the class so that it is more obvious that the
smoke tests are using tempest.manager for it's clients.

Change-Id: Ib4333483c5734d14cac6376837ca770e96bd050e