Remove wait_for_image_resp_code
The nova api first creates an entry in glance,
before responding to snapshot creation request, so the entry
expected to be visible without any additional wait.

This change also introduces an OK code checker on the rest_client side,
it's extended usage can make unnecessary to assert the ok codes in the
test codes.

So the wait_for_image_status will implicitly validate the status code.

Change-Id: I9fdc1cc68501220e6688e2d7a745195b56d60cb9