Create discovery option for stress tests
Adds a decorator @stresstest which automatically sets the type
attr to "stress". It uses the testtools discover functionality to
automatically discover stress tests out of all tempest test.

It is possible to filter stress test with a given attribute type.
For instance to filter out only smoke tests that have a stress test
attribute.

blueprint: stress-tests
Change-Id: I8acf0b608cb500c2679a36a4a00ca4fa14668fad