Always log stdout and stderr of CLI commands
Currently it was logged only when cli subprocess call itself
failed, so there was not enough debug output when asserts/test cases
failed.

Also stderr was logged only when it was merged to the stdout (almost
never).

Now we also don't have any use for check_output method anymore,
as cli.cmd() always captures the stderr,
whatever if for logging or merged into stdout for tests.

Also fixes invalid log.warning on newline at the end of ascii table.

Change-Id: I5251a06018e8bf79c30ad84f8f0ed8aac797f832