Add config options for enabled extensions
This commit adds a new set of config options to the feature_enabled
groups for a list of enabled extensions. These options are used to
specify whether all extensions are enabled or which subset is expected
to be enabled. This just sets up the initial framework for doing
this and converts the FlavorExtraSpecs tests to use it.

paritally implements bp config-cleanup

Change-Id: I6a5a9b16e62eb8a216334a0662c99f0dd0d16873