Add skips to the services decorators
This commit adds a skip to the services decorators. Previously the
decorator was used solely to add an attr to the test_id that specified
which services the test exercised. However, this attr also specifies
which services need to be enabled to run the test. So if a service tag
is set then the decorator will raise a skip exception if the services
specified are not enabled.

Change-Id: Ie5b4f7b7c068b37832e97ffc2a3972b6358425ac