Simplify and move tempest test.py to lib dir
This commit just moves the test.py file from the tempest migration dir
and puts it in the lib dir. As part of the move, everything except for
the base class is split out, for the initial library that's all that is
needed. Other pieces will be added back in over time.