Remove admin param from client method definitions
This commit removes the admin kwarg from the client methods in the
CLIClientBase class. The admin argument was a leftover from tempest
but it doesn't have a place in the new class, because it just takes
in credentials as args. So admin can be specified by just initializing
the class using admin credentials.