Fix migrate_from_tempest script to work with multiple files
The '--follow' option from git log was used to follow the file even if
it was renamed for instance, although the option can't deal with
multiple files which was blocking us to migrate multiple files at once.

This patches fixes the issue by looking one file at time and
concatenating their ids in the end.

Change-Id: I39a2e94e1c27a83de2b976f263928c5199d44b22
Closes-bug: #1411499