Enable stdout passthrough for subunit-trace
One of the biggest issues with subunit-trace is when non-subunit data
is passed into it gets dropped by subunit trace. This patch is an
attempt (based on a similar patch to subunit-2to1) to allow non subunit
data existing before the subunit stream to be printed on stdout. After the
stream is detected however any other stdin unrelated content will still be
lost.

Closes-Bug: #1420067

Change-Id: I94c428120892b987c372473ac2128d73e9ba8f2d