Improve the library's documentation
This commit is the first step in improving the library's documentation.
It adds missing details to a lot of the default cookiecutter docs which
have mostly sat untouched since the repo was created. The release notes
are also brought up to date and adds autodoc pages for all the modules
with public functions.

Change-Id: Id249c95fb269d07628952a4182675bce1fc18a53