Keep original pathes of source files
We are migrating service client code from Tempest, and we need to
migrate files under different pathes in a single patch.
However, current migrating script handles them as the same path and
we could not use the script for the migration.
This patch changes the script for keeping original pathes of source
files.

Change-Id: Id42692aea4b7e132f49dbfe9a0a6551fa2a56770