Change sha1s to change ids in commit msg from migration tool
Recently the migrate_from_tempest.sh script was updated to switch from
grabbing a list of all the sha1s for that file and dumping that into
the commit message to a much cleaner list of Change-Ids for the
most recent commits that touched a file. There was one little
oversight that slipped through that change where we were still
referring to sha1s at the bottom of the commit msg. This commit fixes
that message to now say Change-Ids.

Change-Id: Idd4c7bf469a44ff1758bb2c7d8c61459fc133a2b