Add more debug info for token_client request
Add the request headers and response body to the debug
logs of the token_client request. Request body is omitted because
of security reasons of not revealing passwords in the debug logs.
Also, response body has sufficient information about the type of
the token requested, therefore omitting request headers is ok.

Closes-Bug: #1503751
Change-Id: I830364e1df0e34398039442dc003fe46ca1db3a1